// app/types/theme.ts
export type ThemeProps = {
  theme: {
    colors: {
      background: string;
      primary: string;
      secondary: string;
      gold: string;
      textPrimary: string;
      textSecondary: string;
      textAccent: string;
      success: string;
      warning: string;
      buttonGradient: string[];
    };
  };
};